Provide refactoring

+++ src/service/runtime/src/main/kotlin/io/github/numq/haskcore/service/runtime/LocalRuntimeService.kt
@@ -3,7 +3,7 @@ package io.github.numq.haskcore.service.runtime
 import arrow.core.Either
 import arrow.core.getOrElse
 import arrow.core.right
-import io.github.numq.haskcore.core.timestamp.Timestamp
+import io.github.numq.haskcore.common.core.timestamp.Timestamp
 import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.*
@@ -18,12 +18,18 @@ internal class LocalRuntimeService(private val scope: CoroutineScope) : RuntimeS
 
     override val events = _events.asSharedFlow()
 
+    override suspend fun isActive(id: String) = Either.catch {
+        jobs[id]?.isActive == true
+    }
+
     override suspend fun execute(request: RuntimeRequest): Either<Throwable, Flow<RuntimeEvent>> = channelFlow {
         val startTime = Timestamp.now()
 
         val process = runInterruptible {
             ProcessBuilder(request.command, *request.arguments.toTypedArray()).apply {
-                request.workingDir?.let(::File)?.let(::directory)
+                request.workingDir?.let(::File)?.takeIf(File::exists)?.let(::directory)
+
+                environment().putAll(request.env)
 
                 redirectErrorStream(false)
             }.start()
@@ -49,22 +55,26 @@ internal class LocalRuntimeService(private val scope: CoroutineScope) : RuntimeS
             }
         }
 
-        launch {
-            send(RuntimeEvent.Started(request = request))
+        send(RuntimeEvent.Started(request = request))
 
-            val exitCode = process.waitFor()
+        val exitCode = process.waitFor()
 
-            joinAll(stdoutJob, stderrJob)
+        joinAll(stdoutJob, stderrJob)
 
-            val duration = (Timestamp.now() - startTime).nanoseconds
+        val duration = (Timestamp.now() - startTime).nanoseconds
 
-            send(RuntimeEvent.Terminated(request = request, exitCode = exitCode, duration = duration))
+        val terminatedEvent = RuntimeEvent.Terminated(request = request, exitCode = exitCode, duration = duration)
 
-            close()
-        }
+        trySend(terminatedEvent)
+
+        _events.emit(terminatedEvent)
+
+        close()
 
         awaitClose {
-            process.destroyForcibly()
+            if (process.isAlive) {
+                process.destroyForcibly()
+            }
         }
     }.flowOn(Dispatchers.IO).right()
